fix:change names of types, create ProductRating component
